membros-backend/requirements.txt
asgiref==3.2.10
boto3==1.16.16
botocore==1.19.16
cached-property==1.5.2
Django==3.0.8
django-cors-headers==3.4.0
django-currentuser==0.5.2
django-environ==0.4.5
django-filter==2.4.0
django-storages==1.10.1
django-url-filter==0.3.15
djangorestframework==3.11.0
enum-compat==0.0.3
jmespath==0.10.0
Pillow==7.2.0
psycopg2==2.8.5
python-dateutil==2.8.1
pytz==2020.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.11
watchtower==1.0.0
